backup_old_files/requirements_basic.txt
# IndicAgri Bot - Basic Requirements (Safe Installation)
# This file includes core dependencies that are most likely to install successfully

# Core Python packages
torch>=1.9.0
numpy>=1.21.0
requests>=2.28.0
pyyaml>=6.0

# Web framework
flask>=2.0.0
flask-cors>=4.0.0

# Search capabilities
duckduckgo-search>=3.8.0

# Web scraping
beautifulsoup4>=4.9.3
lxml>=4.6.3

# Data handling
dataclasses-json>=0.5.0
colorlog>=6.0.0

# Machine Learning basics
transformers>=4.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Environment management
python-dotenv>=0.19.0
packaging>=21.0
